Some kind of working

+++ Project/Project/State.h
@@ -16,7 +16,7 @@ public:
 	~State();
 
 	virtual void Initialize() = 0;
-	virtual void Update(float deltatime) = 0;
+	virtual void Update(double deltatime) = 0;
 	virtual void Draw(DrawManager& p_DrawManager) = 0;
 	virtual void Exit() = 0;
